Add BinaryMessageWriter class to allow reuse of XmlWriter
* Also add heurisics for guessing buffer size
